Implemented viewmodel for each view
