--- miniPostApp/miniPost/Modules/Profile/ProfileEntity.swift
@@ -1,2 +1,0 @@
-struct ProfileEntity {
-}
