--- miniPostApp/miniPost/Modules/Profile/ProfileInteractor.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-protocol ProfileInteractorProtocol: class {
-    var presenter: ProfilePresenterProtocol! { get set }
-}
-
-class ProfileInteractor: ProfileInteractorProtocol {
-    weak var presenter: ProfilePresenterProtocol!
-}
