+++ miniPostApp/miniPost/Modules/Users/UsersListView.swift
@@ -1,0 +1,14 @@
+//
+//  UsersListView.swift
+//  miniPost
+//
+//  Created by Amine Hajri on 25.04.24.
+//
+
+import SwiftUI
+
+struct UsersListView: View {
+    var body: some View {
+        UsersListViewBridge()
+    }
+}
